fix: change endpoint for retrieve event reservations
