+++ apps/events/views/__init__.py
@@ -1,3 +1,3 @@
-from .event import EventViewSet, publish_event
+from .event import EventViewSet, publish_event, get_event_reservations
 from .speaker import SpeakerViewSet
-from .reservation import ReservationViewSet, get_event_reservations, check_in
+from .reservation import ReservationViewSet, check_in
