+++ apps/events/views/event.py
@@ -6,6 +6,7 @@ from rest_framework.permissions import AllowAny
 
 from apps.events.models.event import Event
 from apps.events.serializers import CreateEventSerializer, EventSerializer
+from apps.events.serializers.reservation_serializer import ReservationSerializer
 
 
 class EventViewSet(GenericViewSet):
@@ -64,5 +65,29 @@ def publish_event(request, event_id) -> Response:
     event.save()
     return Response(
         {"message": "Event published successfully"},
+        status=status.HTTP_200_OK,
+    )
+
+
+@api_view(["GET"])
+@permission_classes([AllowAny])
+def get_event_reservations(request, event_id) -> Response:
+    """
+    Get all reservations for a specific event.
+    """
+    try:
+        existing_event = Event.objects.get(id=event_id)
+    except:
+        existing_event = None
+
+    if not existing_event:
+        return Response(
+            {"error": "Event not found"},
+            status=status.HTTP_404_NOT_FOUND,
+        )
+
+    reservations = existing_event.reservations.all()
+    return Response(
+        ReservationSerializer(reservations, many=True).data,
         status=status.HTTP_200_OK,
     )
